Setup database with Kotlintest listener.
